python-worker/requirements.txt
redis>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# ML / chemistry
rdkit>=2023.9.1
lightgbm>=4.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.44.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.0.0
